backend/requirements.txt
amqp==5.3.1
anyio==4.4.0
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.0
build==1.2.2.post1
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
channels==3.0.5
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==43.0.1
daphne==3.0.2
Django==5.1.1
django-celery-beat==2.7.0
django-cors-headers==4.4.0
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
hyperlink==21.0.0
idna==3.9
incremental==24.7.2
Jinja2==3.1.4
kombu==5.4.0
MarkupSafe==3.0.2
msgpack==1.1.0
numpy==2.1.1
packaging==24.1
pamqp==3.3.0
pandas==2.2.3
pillow==10.4.0
pip-tools==7.4.1
ply==3.11
prompt_toolkit==3.0.47
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodomex==3.21.0
PyJWT==2.9.0
pyOpenSSL==24.2.1
pyproject_hooks==1.2.0
pysmi==1.5.9
pysnmp-pyasn1==1.1.3
pysnmp-pysmi==1.1.12
pysnmpcrypto==0.0.4
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
redis==5.1.0
requests==2.32.3
requests-toolbelt==1.0.0
service-identity==24.1.0
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
unifi-api==1.0.3
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
wheel==0.45.0
zope.interface==7.0.3
